invert start order for rserver & nginx (not daemon) to avoid 1st start glitch
If nginx is started before Rstudio server, there is a small chance that the first
time the user launch the proxied service, nginx is up but not Rstudio server,
leading to a 502 gateway error (from nginx).

Now rstudio is started first, in background, and nginx is run in foreground mode
(no daemon). There could probably be some glitch if Rstudio server doesn't start
fast enough.